requirements.txt
selenium==4.5.0
loguru==0.5.3
pyTelegramBotAPI==4.7.1
python-dotenv==0.19.2
webdriver_manager==4.0.1